requirements.txt
Django==1.10.6
psycopg2==2.7.1
djangorestframework==3.6.2
